requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.3.4
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
